POM on Certification Page
