+++ Pages/CertificationTab.cs
@@ -1,0 +1,82 @@
+﻿using MARS_QA.Utilities;
+using OpenQA.Selenium;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace MARS_QA.Pages
+{
+    public class CertificationTab
+    {
+       
+            By certificationTab = By.CssSelector("#account-profile-section > div > section:nth-child(3) > div > div > div > div.eight.wide.column > form > div.ui.top.attached.tabular.menu > a:nth-child(4)");
+            By addNewCertification = By.CssSelector("#account-profile-section > div > section:nth-child(3) > div > div > div > div.eight.wide.column > form > div.ui.bottom.attached.tab.segment.tooltip-target.active > div.row > div.twelve.wide.column.scrollTable > div > table > thead > tr > th.right.aligned > div");
+            By addCertificationAward = By.CssSelector("#account-profile-section > div > section:nth-child(3) > div > div > div > div.eight.wide.column > form > div.ui.bottom.attached.tab.segment.tooltip-target.active > div.row > div.twelve.wide.column.scrollTable > div > div > div:nth-child(1) > div > input");
+            By certifiedFrom = By.CssSelector("#account-profile-section > div > section:nth-child(3) > div > div > div > div.eight.wide.column > form > div.ui.bottom.attached.tab.segment.tooltip-target.active > div.row > div.twelve.wide.column.scrollTable > div > div > div:nth-child(2) > div.eight.wide.field > input");
+            By yearDropdown = By.CssSelector("#account-profile-section > div > section:nth-child(3) > div > div > div > div.eight.wide.column > form > div.ui.bottom.attached.tab.segment.tooltip-target.active > div.row > div.twelve.wide.column.scrollTable > div > div > div:nth-child(2) > div.three.wide.field > select");
+            By yearOption = By.CssSelector("#account-profile-section > div > section:nth-child(3) > div > div > div > div.eight.wide.column > form > div.ui.bottom.attached.tab.segment.tooltip-target.active > div.row > div.twelve.wide.column.scrollTable > div > div > div:nth-child(2) > div.three.wide.field > select > option:nth-child(4)");
+            By certificationAddButton = By.CssSelector("#account-profile-section > div > section:nth-child(3) > div > div > div > div.eight.wide.column > form > div.ui.bottom.attached.tab.segment.tooltip-target.active > div.row > div.twelve.wide.column.scrollTable > div > div > div.five.wide.field > input.ui.teal.button");
+
+                
+            //Identify Certification Tab and click it
+            public void CertificationTabM(IWebDriver driver)
+            {
+
+            driver.FindElement(certificationTab).Click();
+
+            }
+
+
+     
+        public void AddNewCertification(IWebDriver driver)
+        {
+            driver.FindElement(addNewCertification).Click();
+            driver.FindElement(addCertificationAward).SendKeys("Diploma");
+            driver.FindElement(certifiedFrom).SendKeys("AUT");
+            driver.FindElement(yearDropdown).Click();
+            driver.FindElement(yearOption).Click();
+            driver.FindElement(certificationAddButton).Click();
+
+        }
+        
+        
+
+        public string GetCertification(IWebDriver driver)
+        {
+            IWebElement actualCertificate = driver.FindElement(By.CssSelector("#account-profile-section > div > section:nth-child(3) > div > div > div > div.eight.wide.column > form > div.ui.bottom.attached.tab.segment.tooltip-target.active > div.row > div.twelve.wide.column.scrollTable > div > table > tbody > tr > td:nth-child(1)"));
+            return actualCertificate.Text;
+
+        }
+
+
+    }
+}
+
+
+// Thread.Sleep(1000);
+
+
+//Identify Add certification box and enter valid certification details
+//   Thread.Sleep(1000);
+
+//Identify and add the certified from box
+//Thread.Sleep(1000);
+
+//Identify Addnew certification button and click
+//   Wait.Waitforclicable(driver, "CssSelector", "#account-profile-section > div > section:nth-child(3) > div > div > div > div.eight.wide.column > form > div.ui.bottom.attached.tab.segment.tooltip-target.active > div.row > div.twelve.wide.column.scrollTable > div > table > thead > tr > th.right.aligned > div");
+
+//Identify and select the year of achievement 
+
+
+//Wait.Waitforclicable(driver, "CssSelector", "#account-profile-section > div > section:nth-child(3) > div > div > div > div.eight.wide.column > form > div.ui.bottom.attached.tab.segment.tooltip-target.active > div.row > div.twelve.wide.column.scrollTable > div > div > div:nth-child(2) > div.three.wide.field > select");
+
+//Wait.Waitforclicable(driver, "CssSelector", "#account-profile-section > div > section:nth-child(3) > div > div > div > div.eight.wide.column > form > div.ui.bottom.attached.tab.segment.tooltip-target.active > div.row > div.twelve.wide.column.scrollTable > div > div > div:nth-child(2) > div.three.wide.field > select > option:nth-child(4)");
+
+
+
+//Identify and click the Certification Add button
+//Wait.Waitforclicable(driver, "CssSelctor", "#account-profile-section > div > section:nth-child(3) > div > div > div > div.eight.wide.column > form > div.ui.bottom.attached.tab.segment.tooltip-target.active > div.row > div.twelve.wide.column.scrollTable > div > div > div.five.wide.field > input.ui.teal.button");
+
